+++ khoros/objects/users.py
@@ -6,7 +6,7 @@
 :Example:           ``khoros.users.create(username='john_doe', email='john.doe@example.com')``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     10 Jan 2022
+:Modified Date:     24 May 2022
 """
 
 import warnings
@@ -1178,3 +1178,16 @@ def get_ids_from_login_list(khoros_object, login_list, return_type='list'):
         id_list.append(user_id)
         id_dict[login] = user_id
     return id_list if return_type == 'list' else id_dict
+
+
+def get_registered_users_count(khoros_object):
+    """This function returns the total count of registered users on the community.
+
+    .. versionadded:: 5.0.0
+
+    :param khoros_object: The core :py:class:`khoros.Khoros` object
+    :type khoros_object: class[khoros.Khoros]
+    :returns: An integer of the total registered users count
+    """
+    response = api.make_v1_request(khoros_object, '/users/registered/count')
+    return response['response']['value']['$']